backend/requirements-base.txt
# All dependencies except PyTorch
absl-py==2.2.2
addict==2.4.0
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
einops==0.8.1
fastapi==0.103.1
filelock==3.18.0
fsspec==2025.3.2
future==1.0.0
grpcio==1.71.0
h11==0.14.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
lmdb==1.6.2
Markdown==3.8
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.4
opencv-python==4.8.0.76
packaging==25.0
Pillow==10.1.0
platformdirs==4.3.7
protobuf==6.30.2
pydantic==2.11.3
pydantic_core==2.33.1
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.13.1
tb-nightly==2.20.0a20250421
tensorboard-data-server==0.7.2
tifffile==2025.3.30
tqdm==4.67.1
typing-inspection==0.4.0
typing-extensions==4.13.2
urllib3==2.4.0
uvicorn==0.23.2
Werkzeug==3.1.3
yapf==0.43.0